flask_app/requirements.txt
dagshub==0.6.3
Flask==3.1.2
joblib==1.3.2
mlflow==3.5.1
mlflow_skinny==3.5.1
mlflow_tracing==3.5.1
nltk==3.9.2
numpy==1.26.4
pandas==2.1.3
prometheus_client==0.23.1
python-dotenv==1.2.1
